Hoist user-context higher in react tree
