+++ packages/vulcan-core/lib/client/components/AppGenerator.jsx
@@ -3,18 +3,20 @@
  */
 import React from 'react';
 import { ApolloProvider } from 'react-apollo';
-import { runCallbacks, Components } from 'meteor/vulcan:lib';
+import { runCallbacks, Components, UserContextWrapper } from 'meteor/vulcan:lib';
 import { CookiesProvider } from 'react-cookie';
 import { BrowserRouter } from 'react-router-dom';
 
 const AppGenerator = ({ apolloClient }) => {
   const App = (
     <ApolloProvider client={apolloClient}>
-        <CookiesProvider>
-            <BrowserRouter>
-                <Components.App />
-            </BrowserRouter>
-        </CookiesProvider>
+    <UserContextWrapper>
+    <CookiesProvider>
+    <BrowserRouter>
+      <Components.App />
+    </BrowserRouter>
+    </CookiesProvider>
+    </UserContextWrapper>
     </ApolloProvider>
   );
   // run user registered callbacks to wrap the app
