+++ packages/vulcan-core/lib/modules/index.js
@@ -22,7 +22,6 @@ export { default as withSingle, useSingle } from './containers/withSingle.js';
 export { default as withCreate } from './containers/withCreate.js';
 export { default as withUpdate, useUpdate } from './containers/withUpdate.js';
 export { default as withDelete } from './containers/withDelete.js';
-export { default as withCurrentUser } from './containers/withCurrentUser.js';
 export { default as withMutation } from './containers/withMutation.js';
 export { default as withSiteData } from './containers/withSiteData.js';
