+++ packages/vulcan-lib/lib/modules/index.js
@@ -28,4 +28,5 @@ export * from './detect_locale.js';
 export * from './graphql_templates.js';
 export * from './validation.js';
 export * from './handleOptions';
+export * from './withCurrentUser.js';
 // export * from './resolvers.js';
